# API/openapi.yaml
openapi: 3.0.3
info:
  title: ReleaseHub API
  version: 1.0.0
  description: Shared HTTP API contract for the ReleaseHub frontend and backend.
servers:
  - url: /
...
paths:
...
  /api/v1/auth/login:
    get:
      operationId: beginAuthLogin
      summary: Start OIDC login
      security: []
      tags: [Auth]
      responses:
        "302":
          description: Redirects to the OIDC provider.
          headers:
            Location:
              schema:
                type: string
                format: uri
        "4XX":
          description: The login configuration or request is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          description: Login could not be started.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/callback:
    get:
      operationId: completeAuthCallback
      summary: Complete the OIDC callback
      security: []
      tags: [Auth]
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: state
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "302":
          description: Creates a BFF Session and redirects to the Web application.
          headers:
            Location:
              schema:
                type: string
        "4XX":
          description: The callback parameters or login state are invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        default:
          description: Callback validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/auth/session:
    get:
      operationId: getAuthSession
      summary: Get the current BFF Session
      tags: [Auth]
      security:
        - sessionCookie: []
      responses:
        "200":
          description: The Session is valid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSessionResponse"
        "401":
          description: The user is unauthenticated or the session has expired.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/logout:
    post:
      operationId: logoutAuthSession
      summary: Revoke the current BFF Session
      tags: [Auth]
      security:
        - sessionCookie: []
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: The local Session was revoked.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogoutResponse"
        "401":
          description: The Session is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/auth/backchannel-logout:
    post:
      operationId: backchannelLogout
      summary: Receive OIDC back-channel logout
      security: []
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/BackchannelLogoutRequest"
      responses:
        "204":
          description: The corresponding local Sessions were revoked.
        "4XX":
          description: The logout token is invalid.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: CSRF validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: releasehub_session
  schemas:
...
    ResponseMeta:
      type: object
      required:
        - requestId
        - timestamp
      properties:
        requestId:
          type: string
          minLength: 1
        timestamp:
          type: string
          format: date-time
...
    AuthSession:
      type: object
      required: [userId, username, mustChangePassword]
      properties:
        userId:
          type: string
          format: uuid
        username:
          type: string
          minLength: 1
        mustChangePassword:
          type: boolean
...
    BackchannelLogoutRequest:
      type: object
      required: [logout_token]
      properties:
        logout_token:
          type: string
          minLength: 1
...
    AuthSessionResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/AuthSession"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
    LogoutResult:
      type: object
      required: [loggedOut]
      properties:
        loggedOut:
          type: boolean
        redirectUrl:
          type: string
          format: uri
    LogoutResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: "#/components/schemas/LogoutResult"
        meta:
          $ref: "#/components/schemas/ResponseMeta"
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
        requestId:
          type: string
          minLength: 1
        details:
          type: object
          additionalProperties: true